notebook/requirements.txt
google-cloud-aiplatform[agent_engines]>=1.126.1
google-adk>=1.18.0
google-cloud-storage>=3.0.0